All logic has been reworked so that "cropped" TIFF files can be used throughout the stitching process for all currently supported output formats.  (Cropped TIFF files are those that are just large enough to contain	the remapped image, and the offset/full size information is stored in the TIFF header).  Modified functions include ReplaceAlphaChannel, CreateAlphaChannels, AddStitchingMasks, FlattenTIFF, createPanorama.  New functions include getCropInformationFromTiff, setFullSizeImageParameters.
Simplified ComputeStitchingMask8bits syntax

CreatePanorama.  Reworked logic to handle intermediate output files and produce final output in various formats.  Added logic that decides when to use cropped TIFF as intermediate format based on chosen output format.  TODO: this could use some tidy up, and it might be a good idea to have cropped TIFF as the default intermediate format for TIFF_m and TIFF_mask, and include a function to "expand" the cropped TIFFs to full size when finished (if the user requests non-cropped TIFF).

Reluctantly commented out calls to TIFFSetWarningHandler and TIFFSetErrorHandler...these cause to GCC to abort, with a series of errors about multiple definition of TIFF functions in libpano and libtiff.

Changed title bar of progress dialog to indicate currently processing image number (indexed at 1) and total images to be processed.